requirements.txt
fastapi
uvicorn
python-dotenv
openai>=1.0.0
python-multipart
agno>=0.2.4
